Revert "Refactor v3 REST endpoint families"
